Move all slot encryption/decryption functions to one place
Eliminates some circular dependencies
